Integrated URL classifiers from github/OWASP/url-classifier to make it easier to vet URLs without resorting to error-prone regexs.
This is experimental.

This change deprecates the old way of white-listing protocols
in the scope of the whole sanitizer and makes available an
AttributePolicyBuilder match method that takes a classifier
so it is easy to scope a URL policy to a particular attribute.

Converting an attribute value to a UrlValue requires a UrlContext.
A UrlContext can encapsulate the base URL of the embedding document,
so should not be scoped to a policy, but rather to a sanitize call.
Therading the context through to Attribute and Element policies
requires an API change to {Attribute,Element}Policy.
This change deprecates those in favor of {Attribute,Element}Policy.V2
which are subinterfaces that take an extra Context object.