Reorganize allowStyling so it either fails or succeeds completely.
A user reported issue involved an IllegalArgumentException.

They may be mistaken.
I can't construct an input which will trigger an
IllegalArgumentException inside allowStyling, but this makes
sure that an exception thrown during CssSchema.union will not
leave an HtmlPolicyBuilder in a state where "style" is globally
allowed but a style filter is not installed.