Update guava.version to 27.1
https://github.com/OWASP/java-html-sanitizer/issues/162 identifies a
problem with depending on older guava versions.

https://ossindex.sonatype.org/vuln/24585a7f-eb6b-4d8d-a2a9-a6f16cc7c1d0
affects Guava < 24.1.

This does not directly affect the sanitizer but could allow denial of
service of clients.

Clients that require older JDKs should use `-Dguava.version=` with `mvn`
to specify a version of their choice or add to their POM's `<properties>`
element

```xml
  <guava.version>123.1</guava.version>
```