Fixed bug: srcset element joining missing space.
The string

    /url1, /url2, /url3

have two invalid URLs since the comma attaches to the URL.

When there is no metadata this should be rendered

    /url1 , /url2 , /url3

In practice this doesn't matter for non-malicious code
since the only URL that has no metadata should be the last,
fallback URL; and there's no attack scenario where appending
a comma to the end of an attacker controlled URL helps the
attacker.

This commit uses " , " to separate URL&metadata pairs in
srcset attributes.

It also adds tests for corner cases of the srcset parser and
valid floating point number parser.